+++ gatsby/src/components/navbar.tsx
@@ -9,10 +9,10 @@ import { isMobile, ResponsiveProps } from "../utils"
 
 import ServicesNavButton from "./nav/ServicesNavButton.tsx"
 import AboutNavButton from "./nav/AboutNavButton.tsx"
+import NavItems from "./nav/NavItems.tsx"
 
 // Menu Icons
 import Chilidogmenu from "../images/svg/chilidogmenu.svg"
-import Menuclose from "../images/svg/menuclose.svg"
 
 import { StyleProps } from "../utils"
 
@@ -20,93 +20,12 @@ import * as stylesMobile from "../styles/navbar/mobile.module.css"
 import * as stylesDesktop from "../styles/navbar/desktop.module.css"
 import * as styles from "../styles/navbar/common.module.css"
 import * as orientationStyles from "../styles/utilities/orientation.module.css"
-import { HorizontalRule } from "./hr.tsx"
-import { hidden, desktop, mobile } from "../styles/utilities/visibility.module.css"
+import { mobile } from "../styles/utilities/visibility.module.css"
 
 import NavDesktop from "./desktop"
 import NavMobile from "./mobile"
 
-interface NavItemInfo {
-  href: string
-  text: string
-}
-
-type NavItemsList = [NavItemInfo, NavItemInfo, NavItemInfo, NavItemInfo, NavItemInfo]
-
-interface NavItemsProps {
-  menuId: string
-  title: string
-  items: NavItemsList
-}
-
-interface ResponsiveProps {
-  mobileMenuActive: boolean
-  setMobileMenuActive: () => void
-}
-
-type ResponsiveNavItemsProps = NavItemsProps | ResponsiveProps
-
 import { SUB_NAV_ICON_SIZE } from "./nav/param.ts"
-
-function NavItems(props: ResponsiveNavItemsProps | StyleProps) {
-  const closeBtnStyle: string = stylesMobile.closeBtn
-  const homeNavItemStyle: string = stylesDesktop.navItemsHomeNav
-  const navItem: string = stylesMobile.mobileNavItem
-
-  function onMenuCloseClick(): void {
-    const menu: HTMLElement | null = document.getElementById(props.menuId)
-    if (menu || !!menu) {
-      menu.classList.remove(stylesMobile.navItemsActive)
-      menu.classList.add(stylesMobile.navItemsHidden)
-      props.setMobileMenuActive(false)
-    }
-  }
-
-  if (!props.mobileMenuActive) {
-    onMenuCloseClick()
-  }
-
-  return (
-    <div id={props.menuId} className={props.className}>
-      <SvgButton
-        className={closeBtnStyle}
-        onClick={onMenuCloseClick}
-        aria-label="Close Navbar"
-        title="Close Navbar"
-        icon={<Menuclose width={40} height={40} data-cy="mobile-menu-close"/>}
-      />
-      <HomeNavButton 
-        dataCy="home-nav-item"
-        className={homeNavItemStyle} text={props.title}
-      />
-      <NavButton
-        dataCy="work-nav-item"
-        className={navItem} href={props.items[0].href} text={props.items[0].text} 
-      />
-      <HorizontalRule className={mobile} />
-      <ServicesNavButton
-        dataCy="services-nav-item"
-        className={navItem} href={props.items[1].href} text={props.items[1].text}
-      />
-      <HorizontalRule className={mobile} />
-      <NavButton
-        dataCy="articles-nav-item"
-        className={navItem} href={props.items[2].href} text={props.items[2].text}
-      />
-      <HorizontalRule className={mobile} />
-      <AboutNavButton
-        dataCy="about-nav-item"
-        className={navItem} href={props.items[5].href} text={props.items[5].text}
-      />
-      <NavButton
-        dataCy="contact-nav-item"
-        className={navItem + " " + desktop}
-        href={props.items[3].href}
-        text={props.items[3].text}
-      />
-    </div>
-  )
-}
 
 export default function NavBar() {
   const [mobileMenuActive, setMobileMenuActive] = React.useState(false)
